Improved log format of TimerMiddleware
Move get_callback_source() up to parent middleware module
